+++ wallet/CHANGELOG.md
@@ -39,6 +39,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   contains a `ParticipationEventRegistrationOptions` and expects a `ParticipationEvents` response;
 - `AccountMethod::GetParticipationEventIds` now also contains a `Node`;
 - Custom `Serialize` implementation for `Error`;
+- Made `TransactionOptionsDto, ReturnStrategy, MintTokenTransactionDto` pub reachable;
 
 ### Removed